+++ resources/scripts/components/servers/overview/ServerOverviewContainer.tsx
@@ -1,5 +1,6 @@
 import ServerAdminBlock from '@/components/servers/overview/ServerAdminBlock'
 import ServerDetailsBlock from '@/components/servers/overview/ServerDetailsBlock'
+import ServerNetworkBlock from '@/components/servers/overview/ServerNetworkBlock'
 import ServerPowerBlock from '@/components/servers/overview/ServerPowerBlock'
 import ServerTerminalBlock from '@/components/servers/overview/ServerTerminalBlock'
 import ServerContentBlock from '@/components/servers/ServerContentBlock'
@@ -13,6 +14,7 @@ const ServerOverviewContainer = () => {
       <div className='grid grid-cols-10 gap-6'>
         <ServerDetailsBlock />
         <ServerTerminalBlock />
+        <ServerNetworkBlock />
         {rootAdmin && <ServerAdminBlock />}
       </div>
     </ServerContentBlock>