+++ CHANGELOG.md
@@ -4,6 +4,17 @@ This file is a running track of new features and fixes to each version of the pa
 
 This project follows [Semantic Versioning](http://semver.org) guidelines.
 
+## v3.6.0-beta
+
+### Fixed
+
+- IPv6 addresses wouldn't display in client area settings
+- Confusing "account password" placeholder when creating a server in the admin area
+
+### Added
+
+- Ability to view IP addresses of a server on the dashboard in the client area
+
 ## v3.5.1-beta
 
 ### Added